@@ -1,81 +1,39 @@
 {
-  "name": "admin-on-rest",
-  "version": "1.0.0",
-  "description": "A frontend Framework for building admin applications on top of REST services, using ES6, React and Material UI",
-  "files": [
-    "*.md",
-    "docs/*.md",
-    "lib",
-    "src"
-  ],
-  "main": "lib/index",
-  "scripts": {
-    "test": "make test"
-  },
-  "authors": [
-    "François Zaninotto"
-  ],
-  "repository": "marmelab/admin-on-rest",
-  "homepage": "https://github.com/marmelab/admin-on-rest#readme",
-  "bugs": "https://github.com/marmelab/admin-on-rest/issues",
-  "license": "MIT",
-  "devDependencies": {
-    "babel-cli": "~6.24.1",
-    "babel-core": "~6.24.1",
-    "babel-eslint": "~7.2.1",
-    "babel-loader": "~6.4.1",
-    "babel-plugin-add-module-exports": "~0.2.1",
-    "babel-plugin-transform-builtin-extend": "~1.1.2",
-    "babel-plugin-transform-react-jsx": "~6.24.1",
-    "babel-plugin-transform-runtime": "~6.23.0",
-    "babel-preset-es2015": "~6.24.1",
-    "babel-preset-react": "~6.24.1",
-    "babel-preset-stage-0": "~6.24.1",
-    "babel-register": "~6.24.1",
-    "chromedriver": "~2.26.1",
-    "css-loader": "~0.28.0",
-    "enzyme": "~2.8.2",
-    "eslint": "~3.19.0",
-    "eslint-config-airbnb": "~14.1.0",
-    "eslint-plugin-import": "~2.2.0",
-    "eslint-plugin-jsx-a11y": "~4.0.0",
-    "eslint-plugin-react": "~6.10.3",
-    "express": "~4.15.2",
-    "extract-text-webpack-plugin": "~2.1.0",
-    "file-api": "~0.10.4",
-    "full-icu": "~1.0.3",
-    "ignore-styles": "~5.0.1",
-    "mocha": "~3.2.0",
-    "react-addons-test-utils": "~15.5.1",
-    "selenium-webdriver": "~3.3.0",
-    "sinon": "~2.1.0",
-    "style-loader": "~0.16.0",
-    "webpack": "~1.13.2",
-    "webpack-dev-server": "~1.16.2"
-  },
-  "dependencies": {
-    "babel-runtime": "~6.18.0",
-    "inflection": "~1.12.0",
-    "lodash.debounce": "~4.0.8",
-    "lodash.defaultsdeep": "~4.6.0",
-    "lodash.get": "~4.4.2",
-    "lodash.set": "~4.3.2",
-    "material-ui": "~0.17.4",
-    "node-polyglot": "2.2.2",
-    "prop-types": "~15.5.7",
-    "query-string": "~4.3.2",
-    "react": "~15.5.4",
-    "react-dom": "~15.5.4",
-    "react-dropzone": "~3.13.1",
-    "react-redux": "~5.0.4",
-    "react-router": "~4.1.0",
-    "react-router-dom": "~4.1.0",
-    "react-router-redux": "~5.0.0-alpha.5",
-    "react-tap-event-plugin": "~2.0.1",
-    "recompose": "~0.23.1",
-    "redux": "~3.6.0",
-    "redux-form": "~6.6.3",
-    "redux-saga": "~0.14.6",
-    "reselect": "~3.0.0"
-  }
+    "private": true,
+    "name": "react-admin",
+    "scripts": {
+        "test": "make test"
+    },
+    "jest": {
+        "setupTestFrameworkScriptFile": "./test-setup.js"
+    },
+    "devDependencies": {
+        "babel-cli": "~6.26.0",
+        "babel-core": "~6.26.0",
+        "babel-eslint": "~7.2.2",
+        "babel-plugin-add-module-exports": "~0.2.1",
+        "babel-plugin-transform-builtin-extend": "~1.1.2",
+        "babel-plugin-transform-react-jsx": "~6.24.1",
+        "babel-plugin-transform-runtime": "~6.23.0",
+        "babel-preset-es2015": "~6.24.1",
+        "babel-preset-react": "~6.24.1",
+        "babel-preset-stage-0": "~6.24.1",
+        "babel-register": "~6.26.0",
+        "cheerio": "~1.0.0-rc.2",
+        "chromedriver": "~2.33.0",
+        "eslint": "~4.6.1",
+        "eslint-config-prettier": "~2.4.0",
+        "eslint-plugin-import": "~2.6.1",
+        "eslint-plugin-prettier": "~2.2.0",
+        "eslint-plugin-react": "~7.3.0",
+        "express": "~4.15.2",
+        "jest": "~21.2.1",
+        "lerna": "~2.5.1",
+        "mocha": "~4.0.1",
+        "prettier": "~1.7.0",
+        "prettier-eslint-cli": "~4.4.0",
+        "raf": "~3.4.0",
+        "selenium-webdriver": "~3.5.0"
+    },
+    "workspaces": ["packages/*"]
 }
