@@ -1,6 +1,4 @@
-const chrome = require('chromedriver');
-const webdriver = require('selenium-webdriver');
+import 'chromedriver';
+import webdriver from 'selenium-webdriver';
 
-module.exports = new webdriver.Builder()
-    .forBrowser('chrome')
-    .build();
+export default new webdriver.Builder().forBrowser('chrome').build();
