@@ -1,11 +1,16 @@
 import { By, until } from 'selenium-webdriver';
 
-module.exports = (url) => (driver) => ({
+export default url => driver => ({
     elements: {
         appLoader: By.css('.app-loader'),
-        input: (name, type = 'input') => By.css(`.create-page ${type}[name='${name}']`),
+        input: (name, type = 'input') =>
+            By.css(`.create-page ${type}[name='${name}']`),
+        inputs: By.css(`.ra-input`),
         submitButton: By.css(".create-page button[type='submit']"),
-        descInput: By.css(".ql-editor"),
+        submitAndAddButton: By.css(
+            ".create-page form>div:last-child button[type='button']"
+        ),
+        descInput: By.css('.ql-editor'),
     },
 
     navigate() {
@@ -19,9 +24,20 @@ module.exports = (url) => (driver) => ({
 
     waitUntilDataLoaded() {
         let continued = true;
-        return driver.wait(until.elementLocated(this.elements.appLoader), 400)
-            .catch(() => continued = false) // no loader - we're on the same page !
-            .then(() => continued ? driver.wait(until.stalenessOf(driver.findElement(this.elements.appLoader))) : true)
+        return driver
+            .wait(until.elementLocated(this.elements.appLoader), 2000)
+            .catch(() => (continued = false)) // no loader - we're on the same page !
+            .then(
+                () =>
+                    continued
+                        ? driver.wait(
+                              until.stalenessOf(
+                                  driver.findElement(this.elements.appLoader)
+                              )
+                          )
+                        : true
+            )
+            .catch(() => {}) // The element might have disapeared before the wait on the previous line
             .then(() => driver.sleep(100)); // let some time to redraw
     },
 
@@ -33,10 +49,10 @@ module.exports = (url) => (driver) => ({
         return input.sendKeys(value);
     },
 
-    setValues(values, clearPreviousValue = true){
+    setValues(values, clearPreviousValue = true) {
         let input;
         let lastPromise;
-        values.forEach((val) => {
+        values.forEach(val => {
             input = driver.findElement(this.elements.input(val.name, val.type));
             if (clearPreviousValue) {
                 input.clear();
@@ -46,12 +62,34 @@ module.exports = (url) => (driver) => ({
         return lastPromise;
     },
 
-    getInputValue(name, type='input'){
-        return driver.findElement(this.elements.input(name, type)).getAttribute('value');
+    getInputValue(name, type = 'input') {
+        return driver
+            .findElement(this.elements.input(name, type))
+            .getAttribute('value');
+    },
+
+    getFields() {
+        return driver.findElements(this.elements.inputs).then(fields =>
+            Promise.all(
+                fields.map(field =>
+                    field.getAttribute('class').then(classes =>
+                        classes
+                            .replace('ra-input-', '')
+                            .replace('ra-input', '')
+                            .trim()
+                    )
+                )
+            )
+        );
     },
 
     submit() {
         driver.findElement(this.elements.submitButton).click();
+        return this.waitUntilDataLoaded();
+    },
+
+    submitAndAdd() {
+        driver.findElement(this.elements.submitAndAddButton).click();
         return this.waitUntilDataLoaded();
     },
 });
