@@ -1,9 +1,11 @@
 import { By, until } from 'selenium-webdriver';
 
-module.exports = (url) => (driver) => ({
+export default url => driver => ({
     elements: {
         appLoader: By.css('.app-loader'),
         input: name => By.css(`.edit-page input[name='${name}']`),
+        inputs: By.css(`.ra-input`),
+        tabs: By.css(`.form-tab`),
         submitButton: By.css(".edit-page button[type='submit']"),
         tab: index => By.css(`button.form-tab:nth-of-type(${index})`),
         title: By.css('.title'),
@@ -20,9 +22,20 @@ module.exports = (url) => (driver) => ({
 
     waitUntilDataLoaded() {
         let continued = true;
-        return driver.wait(until.elementLocated(this.elements.appLoader), 400)
-            .catch(() => continued = false) // no loader - we're on the same page !
-            .then(() => continued ? driver.wait(until.stalenessOf(driver.findElement(this.elements.appLoader))) : true)
+        return driver
+            .wait(until.elementLocated(this.elements.appLoader), 2000)
+            .catch(() => (continued = false)) // no loader - we're on the same page !
+            .then(
+                () =>
+                    continued
+                        ? driver.wait(
+                              until.stalenessOf(
+                                  driver.findElement(this.elements.appLoader)
+                              )
+                          )
+                        : true
+            )
+            .catch(() => {}) // The element might have disapeared before the wait on the previous line
             .then(() => driver.sleep(100)); // let some time to redraw
     },
 
@@ -31,12 +44,39 @@ module.exports = (url) => (driver) => ({
         return input.getAttribute('value');
     },
 
+    getFields() {
+        return driver.findElements(this.elements.inputs).then(fields =>
+            Promise.all(
+                fields.map(field =>
+                    field.getAttribute('class').then(classes =>
+                        classes
+                            .replace('ra-input-', '')
+                            .replace('ra-input', '')
+                            .trim()
+                    )
+                )
+            )
+        );
+    },
+
+    getTabs() {
+        return driver
+            .findElements(this.elements.tabs)
+            .then(tabs => Promise.all(tabs.map(tab => tab.getText())));
+    },
+
     setInputValue(name, value, clearPreviousValue = true) {
         const input = driver.findElement(this.elements.input(name));
         if (clearPreviousValue) {
             input.clear();
         }
         return input.sendKeys(value);
+    },
+
+    clickInput(name) {
+        const input = driver.findElement(this.elements.input(name));
+        input.click();
+        return driver.sleep(200);
     },
 
     gotoTab(index) {
