@@ -4,11 +4,26 @@ import driver from '../chromeDriver';
 
 let listeningServer;
 
-before(() => {
-    const server = express();
-    server.use('/', express.static(path.join(__dirname, '../../example')));
-    listeningServer = server.listen(8083);
-});
+before(
+    () =>
+        new Promise(resolve => {
+            const server = express();
+            server.use(
+                '/',
+                express.static(
+                    path.join(__dirname, '../../packages/ra-example')
+                )
+            );
+            listeningServer = server.listen(8083, resolve);
+        })
+);
+
+before(() =>
+    driver
+        .manage()
+        .window()
+        .setSize(1200, 800)
+);
 
 after(async () => {
     listeningServer.close();
